+++ CHANGES.rst
@@ -499,6 +499,9 @@ astropy.io.misc
 astropy.io.fits
 ^^^^^^^^^^^^^^^
 
+- Implemented skip (after warning) of header cards with reserved
+  keywords in ``table_to_hdu``. [#9390]
+
 astropy.io.registry
 ^^^^^^^^^^^^^^^^^^^